delete confirmation implement

+++ src/pages/MyToys/MyToys.jsx
@@ -1,6 +1,7 @@
 import { useContext, useEffect, useState } from "react";
 import { AuthContext } from "../../providers/AuthProviders/AuthProviders";
 import MyToysRow from "./MyToysRow";
+import Swal from "sweetalert2";
 
 const MyToys = () => {
     const { user } = useContext(AuthContext);
@@ -15,18 +16,40 @@ const MyToys = () => {
 
     // delete toy
     const handleDelete = _id => {
-        fetch(`https://toy-marketplace-server-side-six.vercel.app/deleteToy/${_id}`, {
-            method: "DELETE",
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You want to delete this product!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
         })
-            .then(res => res.json())
-            .then(data => {
-                console.log(data);
-                if (data.deletedCount) {
-
-                    const restToys = toys.filter(toy => toy._id !== _id);
-                    setToys(restToys);
+            .then((result) => {
+                if (result.isConfirmed) {
+                    fetch(`https://toy-marketplace-server-side-six.vercel.app/deleteToy/${_id}`, {
+                        method: "DELETE",
+                    })
+                        .then(res => res.json())
+                        .then(data => {
+                            console.log(data);
+                            if (data.deletedCount) {
+                                Swal.fire(
+                                    'Deleted Successfully!',
+                                    'Your product has been deleted.',
+                                    'success'
+                                )
+                                const restToys = toys.filter(toy => toy._id !== _id);
+                                setToys(restToys);
+                            }
+                        })
                 }
+                // else {
+                //     console.log("cancel");
+                // }
             })
+
+
     };
 
     return (
